sync cli on startup + cleanup
