+++ packages/desktop/src/cli.ts
@@ -1,14 +1,6 @@
 import { invoke } from "@tauri-apps/api/core"
 import { message } from "@tauri-apps/plugin-dialog"
 
-export async function isCliInstalled(): Promise<boolean> {
-  return invoke<boolean>("is_cli_installed")
-}
-
-export async function getInstalledCliPath(): Promise<string | null> {
-  return invoke<string | null>("get_installed_cli_path")
-}
-
 export async function installCli(): Promise<void> {
   try {
     const path = await invoke<string>("install_cli")
